Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is ahead of 'origin/main' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	modified:   core/blog/api/v1/views.py

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Date:      Fri Aug 26 19:25:52 2022 +0430

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   core/blog/api/v1/views.py